+++ contracts/common/lib/serialNumber/OptionSNParser.sol
@@ -13,8 +13,8 @@ library OptionSNParser {
         return address(bytes20(sn << 72));
     }
 
-    function triggerDateOfOpt(bytes32 sn) internal pure returns (uint256) {
-        return uint256(bytes4(sn));
+    function triggerDateOfOpt(bytes32 sn) internal pure returns (uint32) {
+        return uint32(bytes4(sn));
     }
 
     function exerciseDaysOfOpt(bytes32 sn) internal pure returns (uint8) {
